+++ docs/project/changelog.md
@@ -170,6 +170,9 @@ substitutions:
   that are installed via `pyodide.loadPackage` from a custom URL.
   {pr}`2743`
 
+- {{ Fix }} micropip now skips package versions which do not follow PEP440.
+  {pr}`2754`
+
 ### Packages
 
 - {{ Enhancement }} Pillow now supports WEBP image format {pr}`2407`.